+++ changelog.md
@@ -1,22 +1,37 @@
 # Changelog
 
+### 0.6.0 → 0.7.0 (unstable)
+
+Breaking changes:
+
+-   namespaces removed
+-   `TestAgentFactory` now returns a promise
+-   `TesAgentFactoryProps` now requires `Tracer` and doesn't require `AIEngine`
+-   `loadFile` method was removed in favor of `PromptFS`
+
+Other changes
+
+-   interface `TestVoiceAgent` added
+-   interface `Tracer` added
+-   interface `PromptFS` and function `createLocalFS` were added
+
 ### 0.5.0 → 0.6.0 (unstable)
 
 Breaking changes:
 
--   addDirective is removed. Use addMessage with role: 'system' instead.
--   ignoreDirectives → ignoreAddedMessages
+-   `addDirective` is removed. Use `addMessage` with role: 'system' instead.
+-   `ignoreDirectives` → `ignoreAddedMessages`
 
 Other changes:
 
--   AIEngine.sendMessage now accepts "Message" rather than a string
+-   `AIEngine.sendMessage` now accepts `Message` rather than a string
 
 ### 0.4.0 → 0.5.0 (unstable)
 
 Breaking changes:
 
 -   `schema` property replaced with `json` which can be boolean,
--   `setDirectiveFormatter` →
+-   `setDirectiveFormatter` removed
 
 Other changes: